requirements.txt
pywin32>=306
Pillow>=10.0
psutil>=5.9
pystray>=0.19
